Refactor: Reintroduced sections directory:
* moved relevant modules in to sections dir
* excluded some modules from sections dir until
I know what to do with them

Refs: BLAIS5-3651
